rename to FilterProxy
This will differenciate between the WorkFilter and this filter better
